% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNormSig.R
\name{addNormSig}
\alias{addNormSig}
\title{Merges summarySYM with Dunn's Results}
\usage{
addNormSig(anova, summaryData, groupvar, alpha = NULL, keepAll = FALSE)
}
\arguments{
\item{anova}{the anova of data of which significance is to be calculated}

\item{summaryData}{A data frame output by the summarySYM function}

\item{groupvar}{variable to group data by (independent variable)}

\item{alpha}{significance level, default is 0.05 when set to NULL}
}
\value{
The merged file of \code{summaryData} and FSA's dunnTest results
}
\description{
Merges summarySYM with Dunn's Results
}
\examples{
mergeSYM(weightdata = plant_weight_data, winrhizo = root_winrhizo_data, cropname = "Plant Name)
merged_df <- mergeSYM(weightdata = plant_weight_data, winrhizo = root_winrhizo_data, cropname = "Plant Name)
}
\author{
Alex Soupir
}
