% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSYM.R
\name{mergeSYM}
\alias{mergeSYM}
\title{Merges WinRhizo and mass files together}
\usage{
mergeSYM(weightdata, winrhizo, cropname = NULL)
}
\arguments{
\item{weightdata}{A .csv file containing with headers "SYM Jar_Replicate Seed_Replicate Crop Condition Condition Start_Date Observation_Date Weight_g Root_Length_Pix SurfArea"}

\item{winrhizo}{A file containing the raw output of WinRhizo converted converted to .csv format}

\item{cropname}{A name that will be added to the output file along with the date that the function was run}
}
\value{
The merged file of \code{weightdata} and \code{winrhizo}
}
\description{
Merges WinRhizo and mass files together
}
\examples{
mergeSYM(weightdata = plant_weight_data, winrhizo = root_winrhizo_data, cropname = "Crop title")

}
