% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perChange.R
\name{perChange}
\alias{perChange}
\title{Adds another column to the table output by summarySYM with the percent change}
\usage{
perChange(data = NULL, perChange = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{The table that was output by summarySYM to add a percent change column to}

\item{perChange}{The row label that is to be compared to}

\item{na.rm}{Set true to not include cells with missing values}
}
\value{
Table with new column of percent change about \code{perChange}
}
\description{
Adds another column to the table output by summarySYM with the percent change
}
\examples{
perChange(summary_table, perChange="Control", na.rm=TRUE)
new.summary_table <- perChange(summary_table, perChange="Control")
}
