% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConf.R
\name{setConf}
\alias{setConf}
\title{Set confidence interval to be used when running summarySYM()}
\usage{
setConf()
}
\value{
temp variable that holds confidence interval for summarySYM()
}
\description{
Set confidence interval to be used when running summarySYM()
}
\examples{
setConf()

}
