% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySYM.R
\name{summarySYM}
\alias{summarySYM}
\title{Creates a summary table for a metric that is set using measurevar}
\usage{
summarySYM(data = NULL, measurevar, change = NULL, groupvars = NULL,
  na.rm = FALSE, conf.interval = temp, .drop = TRUE)
}
\arguments{
\item{data}{A dataset that has raw data to be summarized}

\item{measurevar}{A column heading that is to be summarized}

\item{change}{The row label that is to be used to compare to (control)}

\item{groupvars}{The column heading(s) for summary (isolate)}

\item{na.rm}{Set true to not include cells with missing values}
}
\value{
Table with summarized information about \code{measurevar} sorted by \code{groupvars}
}
\description{
Creates a summary table for a metric that is set using measurevar
}
\examples{
summarySYM(plant_merged, measurevar="Root_Length", groupvars=c("Isolate","Nutrient"), na.rm=TRUE)
summary_table <- summarySYM(plant_merged, measurevar="Root_Length", groupvars="Isolate", na.rm=TRUE)
}
